+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 All notable changes to the "prettier-vscode" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
 ## [Unreleased]
-- New setting `jsxBracketSameLine`.
+- New setting `jsxBracketSameLine`. (prettier 0.17.0)
+- Changed `trailingComma` setting `['none', 'es5', 'all']` (prettier 0.19.0)
 
 ## [0.7.0]
 - Removed `Prettier` action.